last version with doc

+++ DE_Files/twitch-import/push_to_s3.py
@@ -5,22 +5,40 @@ import boto3
 from botocore.exceptions import ClientError
 from time import time, sleep
 
-# define keys for S3 bucket
-access_key = 'AKIAUSHRZ5XVFRNBEMFP'
-access_secret = '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'
-bucket_name = 'twitch-logs'
-LOGDIR = 'log'
-duration = 60
+###############################################################################
+# push_to_s3.py script
+# it connect to an S3 bucket and upload all files present in LOGDIR
+# except if they are named *.log , to avoid uploading an incomplete logfile
+# it remove each file uploaded
+# wait 30 seconds for a new set of logfiles 
+# until reaching the duration parameter (identical in the 2 scripts)
+###############################################################################
 
+# Retreive keys for S3 bucket
+access_key = os.environ['access_key']
+access_secret = os.environ['access_secret']
+access_secret = os.environ['bucket_name']
+
+# define log directory to look at and how long in seconds
+LOGDIR = 'log'
+duration = 300
 
 def connect_to_s3():
+
+    '''
+    Function to connect to the dedicated S3 bucket
+    
+    Parameters : None, they are environment variables
+
+    Returns : an client s3 object
+    '''
 
     try:
         # Connect to S3 bucket by key
     
         s3 = boto3.client(
             's3',
-            region_name="eu-west-1",
+            region_name="eu-north-1",
             aws_access_key_id = access_key,
             aws_secret_access_key = access_secret
         )
@@ -31,37 +49,60 @@ def connect_to_s3():
         print(e)
 
 def push(client_s3, data_file_folder) :  
-    # push all files in LOGDIR to th S3 bucket
-    # and remove them from local if success
+    '''
+    Function to push all files in LOGDIR to the S3 bucket except *.log files 
+    and remove them from local if success
 
+    Parameters :
+        client_s3 : a boto3.client object 
+        data_file_folder : folder in full path to look at
+    
+    Returns : 
+        print a comment when a file is upload or delete
+    '''
+
+    # look at data_file_folder directory
     for file in os.listdir(data_file_folder):
+        
+        # for each file not finishing by .log try to upload it
+        if not file.endswith('log'):
+            try:
+                print('Uploading file {0} ...'.format(file))
+                
+                client_s3.upload_file(
+                    os.path.join(data_file_folder, file),
+                    bucket_name,
+                    file
+                    )
 
-        try:
-            print('Uploading file {0} ...'.format(file))
-            
-            client_s3.upload_file(
-                os.path.join(data_file_folder, file),
-                bucket_name,
-                file
-                )
+                # if it is OK let's remove the local file
+                print('Removing local file {0} ...'.format(file))
+                os.remove(os.path.join(data_file_folder, file))
 
-            # if it is OK let's remove the local file
-            print('Removing local file {0} ...'.format(file))
-            os.remove(os.path.join(data_file_folder, file))
+            except ClientError as e:
+                print('Incorrect credential')
+                print(e)
 
-        except ClientError as e:
-            print('Incorrect credential')
-            print(e)
-
-        except Exception as e:
-            print(e)    
+            except Exception as e:
+                print(e)    
 
 
 def main():
+
+    '''
+    Main function to define in full path the logfile directory 
+    open the connection to s3 blucket and loop on push function 
+    on a 30 second time period (same as logfiles creation)
+
+    Parameters : None all is in environment 
+
+    Returns : None
+
+    '''
     
     client_s3 = connect_to_s3()
 
-    # repare path , better to be in full path
+    # Prepare path , better to be in full path
     data_file_folder = os.path.join(os.getcwd(), LOGDIR)
 
     startTime = time()
@@ -73,6 +114,9 @@ def main():
 
 
 def init():
+    '''
+    init Function to start main() and exit properly
+    '''
   if __name__ == '__main__':
     sys.exit(main())
